SNAP-715: reduce result set columns
